Extract fullname mention keys logic to reusable custom hook
This commit:
1. Creates a new useFullnameMentionKeys custom hook
2. Simplifies the implementation to use a single loop for all users
3. Updates advanced_text_editor.tsx to use the new hook
4. Improves code maintainability and reduces duplication
